requirements.txt
allure-pytest==2.8.5
allure-python-commons==2.8.5
psycopg2==2.8.3
pyparsing==2.4.2
pytest==5.1.2
requests==2.22.0
urllib3==1.25.3
